move math intrinsics to their own file
